refactor: refactored project's architecture
