+++ src/modules/user-activities-answered/services/user-activities-answered.service.spec.ts
@@ -2,14 +2,14 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserActivitiesAnsweredService } from './user-activities-answered.service';
 import { DatabaseModule } from '../../../database/database.module';
-import { UserModule } from '../../user/user.module';
+import { UserModule } from '../../user/infra/modules/user.module';
 import { User } from '../../user/entity/user.entity';
 import { UserScore } from '../../user-score/entity/user-score.entity';
 import { UserScoreService } from '../../user-score/services/user-score.service';
 import { UserScoreRepository } from '../../user-score/repository/user-score.repository';
-import { UserRepository } from '../../user/repository/user.repository';
+import { UserRepository } from '../../user/repositories/user.repository';
 import { JWTProvider } from '../../user/providers/jwt.provider';
-import { UserNotFoundException } from '../../user/domain/errors/UserNotFound.exception';
+import { UserNotFoundException } from '../../user/domain/dtos/errors/UserNotFound.exception';
 import { ActivityNotFoundException } from '../domain/errors/ActivityNotFound.exception';
 import { ActivityAlreadyAnsweredException } from '../domain/errors/ActivityAlreadyAnswered.exception';
 import { WrongAnswerException } from '../domain/errors/WrongAnswer.exception';
