+++ src/modules/user-score/controller/user-score.controller.ts
@@ -5,7 +5,7 @@ import { NotAuthenticatedException } from '../../../shared/domain/errors/NotAuth
 import { AllExceptionsFilterDTO } from '../../../shared/domain/dtos/errors/AllException.filter.dto';
 import { CommonException } from '../../../shared/domain/errors/Common.exception';
 import { Request, Response } from 'express';
-import { MultipleTopScoresResponseDTO } from '../../user/domain/requests/FindTopScores.request.dto';
+import { MultipleTopScoresResponseDTO } from '../../user/domain/dtos/requests/FindTopScores.request.dto';
 
 @Controller('scores')
 @ApiTags('Placar de Usuários')
