+++ src/modules/user-score/services/user-score.service.spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UserScoreService } from './user-score.service';
 import { DatabaseModule } from '../../../database/database.module';
-import { UserModule } from '../../user/user.module';
+import { UserModule } from '../../user/infra/modules/user.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../../user/entity/user.entity';
 import { UserScore } from '../entity/user-score.entity';
 import { UserScoreRepository } from '../repository/user-score.repository';
-import { UserRepository } from '../../user/repository/user.repository';
+import { UserRepository } from '../../user/repositories/user.repository';
 import { JWTProvider } from '../../user/providers/jwt.provider';
-import { UserNotFoundException } from '../../user/domain/errors/UserNotFound.exception';
+import { UserNotFoundException } from '../../user/domain/dtos/errors/UserNotFound.exception';
 import { UnprocessableDataException } from '../../../shared/domain/errors/UnprocessableData.exception';
 import { TestHelper } from '../../../../test/helpers/dbInstanceHelper';
