+++ src/modules/user-score/services/user-score.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { UserScoreRepository } from '../repository/user-score.repository';
 import { UnprocessableDataException } from '../../../shared/domain/errors/UnprocessableData.exception';
-import { UserNotFoundException } from '../../user/domain/errors/UserNotFound.exception';
-import { FindTopScoresResponseDTO } from '../../user/domain/requests/FindTopScores.request.dto';
-import { UserRepository } from '../../user/repository/user.repository';
+import { UserNotFoundException } from '../../user/domain/dtos/errors/UserNotFound.exception';
+import { FindTopScoresResponseDTO } from '../../user/domain/dtos/requests/FindTopScores.request.dto';
+import { UserRepository } from '../../user/repositories/user.repository';
 
 @Injectable()
 export class UserScoreService {
