+++ src/modules/user-score/user-score.module.ts
@@ -12,11 +12,11 @@ import {
 import { User } from '../user/entity/user.entity';
 import { DatabaseModule } from 'src/database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserModule } from '../user/user.module';
+import { UserModule } from '../user/infra/modules/user.module';
 import { UserScoreRepository } from './repository/user-score.repository';
 import { UserScore } from './entity/user-score.entity';
-import { UserRepository } from '../user/repository/user.repository';
-import { AuthenticationMiddleware } from '../user/middlewares/auth.middleware';
+import { UserRepository } from '../user/repositories/user.repository';
+import { AuthenticationMiddleware } from '../user/infra/http/middlewares/auth.middleware';
 import { JWTProvider } from '../user/providers/jwt.provider';
 
 @Module({
