+++ src/modules/user/domain/dtos/errors/EmailAlreadyRegistered.exception.ts
