+++ src/modules/user/domain/dtos/errors/UserNotFound.exception.ts
