+++ src/modules/user/domain/dtos/errors/UsernameAlreadyRegistered.exception.ts
